requirements.txt
anthropic>=0.39.0
streamlit>=1.40.0
streamlit-shadcn-ui>=0.1.19
pydantic>=2.0.0
python-dotenv>=1.0.0
